Add label to custom resource

+++ pkg/controller/operandrequest/reconcile_operand.go
@@ -42,6 +42,7 @@ func (r *ReconcileOperandRequest) reconcileOperand(requestInstance *operatorv1al
 		for _, operand := range req.Operands {
 			configInstance, err := r.getConfigInstance(req.Registry, req.RegistryNamespace)
 			if err != nil {
+				klog.Error(err)
 				merr.Add(err)
 				continue
 			}
@@ -54,6 +55,7 @@ func (r *ReconcileOperandRequest) reconcileOperand(requestInstance *operatorv1al
 
 				// If can't get CSV, requeue the request
 				if err != nil {
+					klog.Error(err)
 					merr.Add(err)
 					continue
 				}
@@ -139,12 +141,15 @@ func (r *ReconcileOperandRequest) createUpdateCr(service *operatorv1alpha1.Confi
 
 		// Get the kind of CR
 		kind := unstruct.Object["kind"].(string)
+		apiversion := unstruct.Object["apiVersion"].(string)
+		name := unstruct.Object["metadata"].(map[string]interface{})["name"].(string)
 
+		var found bool
 		for crdName, crConfig := range service.Spec {
-
 			// Compare the name of OperandConfig and CRD name
 			if strings.EqualFold(kind, crdName) {
 				klog.V(4).Info("Found OperandConfig spec for custom resource: " + kind)
+				found = true
 				//Convert CR template spec to string
 				specJSONString, _ := json.Marshal(unstruct.Object["spec"])
 
@@ -153,6 +158,11 @@ func (r *ReconcileOperandRequest) createUpdateCr(service *operatorv1alpha1.Confi
 
 				unstruct.Object["spec"] = mergedCR
 				unstruct.Object["metadata"].(map[string]interface{})["namespace"] = namespace
+				if (unstruct.Object["metadata"].(map[string]interface{})["labels"] == nil) {
+					klog.V(4).Info("Adding ODLM label in the custom resource")
+					unstruct.Object["metadata"].(map[string]interface{})["labels"] = make(map[string]interface{})
+				}
+				unstruct.Object["metadata"].(map[string]interface{})["labels"].(map[string]interface{})["operator.ibm.com/opreq-control"] = "true"
 
 				// Creat or Update the CR
 				crCreateErr := r.client.Create(context.TODO(), &unstruct)
@@ -167,13 +177,13 @@ func (r *ReconcileOperandRequest) createUpdateCr(service *operatorv1alpha1.Confi
 				} else if errors.IsAlreadyExists(crCreateErr) {
 					existingCR := &unstructured.Unstructured{
 						Object: map[string]interface{}{
-							"apiVersion": unstruct.Object["apiVersion"].(string),
-							"kind":       unstruct.Object["kind"].(string),
+							"apiVersion": apiversion,
+							"kind":       kind,
 						},
 					}
 
 					crGetErr := r.client.Get(context.TODO(), types.NamespacedName{
-						Name:      unstruct.Object["metadata"].(map[string]interface{})["name"].(string),
+						Name:      name,
 						Namespace: namespace,
 					}, existingCR)
 
@@ -201,7 +211,6 @@ func (r *ReconcileOperandRequest) createUpdateCr(service *operatorv1alpha1.Confi
 					if stateUpdateErr != nil {
 						merr.Add(stateUpdateErr)
 					}
-
 				} else {
 					klog.V(2).Info("Finish creating the Custom Resource: ", crdName)
 					stateUpdateErr := r.updateServiceStatus(csc, service.Name, crdName, operatorv1alpha1.ServiceRunning)
@@ -211,12 +220,42 @@ func (r *ReconcileOperandRequest) createUpdateCr(service *operatorv1alpha1.Confi
 				}
 			}
 		}
+		if (!found) {
+			crShouldBeDeleted := &unstructured.Unstructured{
+				Object: map[string]interface{}{
+					"apiVersion": apiversion,
+					"kind":       kind,
+				},
+			}
+			getError := r.client.Get(context.TODO(),types.NamespacedName{
+				Name: name,
+				Namespace: namespace,
+			}, crShouldBeDeleted)
+			if getError != nil && !errors.IsNotFound(getError) {
+				klog.Error(getError)
+				merr.Add(getError)
+				continue
+			}
+			if !errors.IsNotFound(getError) && crShouldBeDeleted.Object["metadata"].(map[string]interface{})["labels"].(map[string]interface{})["operator.ibm.com/opreq-control"] == "true"  {
+				klog.V(3).Infof("Deleting custom resource: %s from custom resource definition: %s", name, kind)
+				deleteErr := r.client.Delete(context.TODO(),crShouldBeDeleted)
+				if deleteErr != nil {
+					klog.Error(deleteErr)
+					merr.Add(deleteErr)
+					continue
+				}
+				stateDeleteErr := r.deleteServiceStatus(csc, service.Name, util.Lcfirst(kind))
+				if stateDeleteErr != nil {
+					klog.Error("Failed to clean up the deleted service status in the operand config: ", stateDeleteErr)
+					merr.Add(stateDeleteErr)
+					continue
+				}
+				klog.V(3).Infof("Finish deleting custom resource: %s from custom resource definition: %s", name, kind)			}
+		}
 	}
-
 	if len(merr.errors) != 0 {
 		return merr
 	}
-
 	return nil
 }
 
@@ -256,13 +295,16 @@ func (r *ReconcileOperandRequest) deleteCr(csv *olmv1alpha1.ClusterServiceVersio
 
 			// Compare the name of OperandConfig and CRD name
 			if strings.EqualFold(kind, crdName) {
-				crDeleteErr := r.client.Delete(context.TODO(), &unstruct)
-				if crDeleteErr != nil && !errors.IsNotFound(crDeleteErr) {
-					klog.Error("Failed to delete the custom resource: ", crDeleteErr)
-					merr.Add(crDeleteErr)
+				getError := r.client.Get(context.TODO(),types.NamespacedName{
+					Name: name,
+					Namespace: namespace,
+				},  &unstruct)
+				if getError != nil && !errors.IsNotFound(getError) {
+					klog.Error(getError)
+					merr.Add(getError)
 					continue
 				}
-				if errors.IsNotFound(crDeleteErr) {
+				if errors.IsNotFound(getError) {
 					klog.V(4).Info("Deleted the CR: " + kind)
 					stateDeleteErr := r.deleteServiceStatus(csc, service.Name, crdName)
 					if stateDeleteErr != nil {
@@ -271,35 +313,43 @@ func (r *ReconcileOperandRequest) deleteCr(csv *olmv1alpha1.ClusterServiceVersio
 					}
 					continue
 				}
-				klog.V(4).Info("Waiting for CR: " + kind + " is deleted")
-				err := wait.PollImmediate(time.Second*20, time.Minute*10, func() (bool, error) {
-					klog.V(4).Info("Checking for CR: " + kind + " is deleted")
-					err := r.client.Get(context.TODO(), types.NamespacedName{
-						Name:      name,
-						Namespace: namespace,
-					},
-						&unstruct)
-					if errors.IsNotFound(err) {
-						return true, nil
+				if unstruct.Object["metadata"].(map[string]interface{})["labels"].(map[string]interface{})["operator.ibm.com/opreq-control"] == "true" {
+					crDeleteErr := r.client.Delete(context.TODO(), &unstruct)
+					if crDeleteErr != nil && !errors.IsNotFound(crDeleteErr) {
+						klog.Error("Failed to delete the custom resource: ", crDeleteErr)
+						merr.Add(crDeleteErr)
+						continue
 					}
+					klog.V(4).Info("Waiting for CR: " + kind + " is deleted")
+					err := wait.PollImmediate(time.Second*20, time.Minute*10, func() (bool, error) {
+						klog.V(4).Info("Checking for CR: " + kind + " is deleted")
+						err := r.client.Get(context.TODO(), types.NamespacedName{
+							Name:      name,
+							Namespace: namespace,
+						},
+							&unstruct)
+						if errors.IsNotFound(err) {
+							return true, nil
+						}
+						if err != nil {
+							klog.Error("Failed to get the custom resource: ", err)
+							return false, err
+						}
+						return false, nil
+					})
 					if err != nil {
-						klog.Error("Failed to get the custom resource: ", err)
-						return false, err
+						klog.Error(err)
+						merr.Add(err)
+						continue
 					}
-					return false, nil
-				})
-				if err != nil {
-					klog.Error(err)
-					merr.Add(err)
-					continue
+					stateDeleteErr := r.deleteServiceStatus(csc, service.Name, crdName)
+					if stateDeleteErr != nil {
+						klog.Error("Failed to clean up the deleted service status in the operand config: ", stateDeleteErr)
+						merr.Add(stateDeleteErr)
+						continue
+					}
+					klog.V(4).Info("Deleted the CR: " + kind)
 				}
-				stateDeleteErr := r.deleteServiceStatus(csc, service.Name, crdName)
-				if stateDeleteErr != nil {
-					klog.Error("Failed to clean up the deleted service status in the operand config: ", stateDeleteErr)
-					merr.Add(stateDeleteErr)
-					continue
-				}
-				klog.V(4).Info("Deleted the CR: " + kind)
 			}
 
 		}
@@ -313,7 +363,7 @@ func (r *ReconcileOperandRequest) deleteCr(csv *olmv1alpha1.ClusterServiceVersio
 
 // Get the OperandConfig instance with the name and namespace
 func (r *ReconcileOperandRequest) getConfigInstance(name, namespace string) (*operatorv1alpha1.OperandConfig, error) {
-	klog.V(4).Info("Get the OperandConfig instance from the name: ", name, " namespace: ", namespace)
+	klog.V(3).Info("Get the OperandConfig instance from the name: ", name, " namespace: ", namespace)
 	config := &operatorv1alpha1.OperandConfig{}
 	if err := r.client.Get(context.TODO(), types.NamespacedName{Name: name, Namespace: namespace}, config); err != nil {
 		return nil, err
